+++ common/src/main/java/com/st0x0ef/stellaris/client/screens/tablet/TabletEntryScreen.java
@@ -1,7 +1,6 @@
 package com.st0x0ef.stellaris.client.screens.tablet;
 
 import com.mojang.blaze3d.systems.RenderSystem;
-import com.st0x0ef.stellaris.Stellaris;
 import com.st0x0ef.stellaris.client.screens.components.TabletButton;
 import com.st0x0ef.stellaris.client.screens.components.TexturedButton;
 import com.st0x0ef.stellaris.common.utils.ResourceLocationUtils;