+++ changelog.md
@@ -4,6 +4,7 @@
 - Add diesel
 - Add diesel generator
 - Add new config system
+- Add power banks
 
 ## Changes
 - Rover now need diesel instead of fuel